english/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
Django==4.2.1
django-extensions==3.2.3
django-shell-plus==1.1.7
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.3.0
